Considerar -> Consider
Considerar appears to be spanish.  I believe "Consider" is appropriate here, although it's still not a hugely informative tooltip.  That said, it's not a hugely complicated checkbox, either.

Someone who knows the project better might want to consider changing it to "Consider hidden nodes in searches" if that intent is correct.